backend/requirements.txt
fastapi
uvicorn
motor
beanie
pydantic-settings
python-dotenv
email-validator
passlib[bcrypt]
python-jose[cryptography]
python-multipart
